signer: move signer impl to key to abstract Signer() call
