+++ agent/agent.go
@@ -20,7 +20,6 @@ import (
 	keyfile "github.com/foxboron/go-tpm-keyfiles"
 	"github.com/foxboron/ssh-tpm-agent/internal/keyring"
 	"github.com/foxboron/ssh-tpm-agent/key"
-	"github.com/foxboron/ssh-tpm-agent/signer"
 	"github.com/google/go-tpm/tpm2/transport"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
@@ -107,12 +106,13 @@ func (a *Agent) signers() ([]ssh.Signer, error) {
 	}
 
 	for _, k := range a.keys {
-		s, err := ssh.NewSignerFromSigner(
-			signer.NewSSHKeySigner(k, a.keyring(), a.op, a.tpm,
-				func(_ *keyfile.TPMKey) ([]byte, error) {
-					// Shimming the function to get the correct type
-					return a.pin(k)
-				}))
+		s, err := ssh.NewSignerFromSigner(k.Signer(
+			a.keyring(), a.op, a.tpm,
+			func(_ *keyfile.TPMKey) ([]byte, error) {
+				// Shimming the function to get the correct type
+				return a.pin(k)
+			}),
+		)
 		if err != nil {
 			return nil, fmt.Errorf("failed to prepare signer: %w", err)
 		}
